Fixed parsing to float/int issue

+++ ParseFunctions.js
@@ -23,8 +23,8 @@ class ParseFunctions {
 
             const newTradeObject =
                 {
-                    price: parseInt(price),
-                    size: parseInt(size)
+                    price: parseFloat(price),
+                    size: parseFloat(size)
                 };
 
             arrayOfParsedBids.push(newTradeObject);
@@ -38,8 +38,8 @@ class ParseFunctions {
 
             const newTradeObject =
                 {
-                    price: parseInt(price),
-                    size: parseInt(size)
+                    price: parseFloat(price),
+                    size: parseFloat(size)
                 };
 
             arrayOfParsedAsks.push(newTradeObject);
